Set a timeout for finish and login expect script.
Running scripts without timeout leads to stalled processes without
progress.  It is better to fail with error and try again.  Especially
jobs run by cron get a chance to recover this way.

Change the finish and login timeout from infinite to 3500 seconds.
This is a bit less than an hour as the console program prints a
mark every hour.  This log should not interrupt the timeout.